Support symlinking transform.sh into a different directory
I like to symlink scripts in ~/.local/bin so I can call them from any
directory and not have to deal with long paths. However, the script does
not work properly if `from_to_flat.txt` and `from_to.txt` are not in the
same directory as `transform.sh`. This PR fixes that problem by first
passing `$0` to `readlink -f` *and then* passing that output to
`dirname` in order to figure out where those files are located.

Signed-off-by: Maxwell G <gotmax@e.email>